Time: 3 ms (47.21%), Space: 5.8 MB (96.11%) - LeetHub

+++ 50-powx-n/50-powx-n.cpp
@@ -10,7 +10,7 @@ public:
             nn=-nn;
         }
         
-        long double ans = myPow(x, nn/2);
+        double ans = myPow(x, nn/2);
         if(nn%2==0) ans = ans * ans;
         else  ans = ans * ans * x;
